feat: Add Smart Card Emulation protocol methods and types
- Implemented methods for enabling, disabling, and reporting results for Smart Card Emulation.
- Added corresponding types for arguments and results in smartcardemulation/types.go.

feat: Enhance Storage protocol with new tracking and bucket management methods

- Introduced methods for tracking storage buckets and managing bounce tracking mitigations.
- Updated types to support new functionalities in storage/types.go.

feat: Update System Info protocol to use 'any' type for GPU attributes

- Modified GPUInfo struct to use 'any' for AuxAttributes and FeatureStatus in systeminfo/types.go.

fix: Correct documentation typos in Target protocol methods

- Fixed typos in comments for SetAutoAttach and other methods in target/methods.go.

feat: Add WebMCP protocol for tool management

- Created WebMCP protocol with methods for enabling/disabling the domain and invoking tools.
- Defined events for tool addition, removal, invocation, and response in webmcp/events.go.

feat: Extend WebAuthn protocol with credential management features

- Added methods for setting credential properties and handling credential updates in webauthn/methods.go.
- Updated types to include new fields for backup eligibility and state in webauthn/types.go.

fix: Update session initialization to enable file chooser events

- Modified session.go to enable file chooser opened events during page initialization.

+++ protocol/input/types.go
@@ -14,8 +14,8 @@ type TouchPoint struct {
 	RotationAngle      float64 `json:"rotationAngle,omitempty"`
 	Force              float64 `json:"force,omitempty"`
 	TangentialPressure float64 `json:"tangentialPressure,omitempty"`
-	TiltX              int     `json:"tiltX,omitempty"`
-	TiltY              int     `json:"tiltY,omitempty"`
+	TiltX              float64 `json:"tiltX,omitempty"`
+	TiltY              float64 `json:"tiltY,omitempty"`
 	Twist              int     `json:"twist,omitempty"`
 	Id                 float64 `json:"id,omitempty"`
 }
@@ -99,8 +99,8 @@ type DispatchMouseEventArgs struct {
 	ClickCount         int                   `json:"clickCount,omitempty"`
 	Force              float64               `json:"force,omitempty"`
 	TangentialPressure float64               `json:"tangentialPressure,omitempty"`
-	TiltX              int                   `json:"tiltX,omitempty"`
-	TiltY              int                   `json:"tiltY,omitempty"`
+	TiltX              float64               `json:"tiltX,omitempty"`
+	TiltY              float64               `json:"tiltY,omitempty"`
 	Twist              int                   `json:"twist,omitempty"`
 	DeltaX             float64               `json:"deltaX,omitempty"`
 	DeltaY             float64               `json:"deltaY,omitempty"`
